week 2

+++ 2- Week Home Work/src/repository/CrudRepository.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public interface CrudRepository<T> {
     List<T> listall();
-     T finfById(int id);
+     T finfById(Long id);
      void saveToDatabase(T t);
-     void updateOnDatabase(T t ,int id);
+     void updateOnDatabase(T t ,Long id);
      void deletFromDatabase(T t);
-     void deletFromDatabase(T t ,int id);
+     void deletFromDatabase(T t ,Long id);
